+++ packages/web3-eth/src/utils/format_transaction.ts
@@ -18,7 +18,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import { Transaction, DataFormat, DEFAULT_RETURN_FORMAT, FormatType } from 'web3-types';
 import { isNullish, ValidationSchemaInput } from 'web3-validator';
 import { TransactionDataAndInputError } from 'web3-errors';
-import { mergeDeep, bytesToBuffer, format } from 'web3-utils';
+import { mergeDeep, bytesToHex, format } from 'web3-utils';
 import { transactionSchema } from '../schemas';
 
 export function formatTransaction<
@@ -46,8 +46,8 @@ export function formatTransaction<
 		formattedTransaction.data !== formattedTransaction.input
 	)
 		throw new TransactionDataAndInputError({
-			data: bytesToBuffer(formattedTransaction.data).toString('hex'),
-			input: bytesToBuffer(formattedTransaction.input).toString('hex'),
+			data: bytesToHex(formattedTransaction.data),
+			input: bytesToHex(formattedTransaction.input),
 		});
 	else if (!isNullish(formattedTransaction.data)) {
 		formattedTransaction.input = formattedTransaction.data;