+++ packages/web3-eth/test/unit/rpc_method_wrappers/fixtures/return_formats.ts
@@ -22,10 +22,10 @@ export const returnFormats: DataFormat[] = [
 	{ number: FMT_NUMBER.STR, bytes: FMT_BYTES.HEX },
 	{ number: FMT_NUMBER.BIGINT, bytes: FMT_BYTES.HEX },
 
-	{ number: FMT_NUMBER.NUMBER, bytes: FMT_BYTES.BUFFER },
-	{ number: FMT_NUMBER.HEX, bytes: FMT_BYTES.BUFFER },
-	{ number: FMT_NUMBER.STR, bytes: FMT_BYTES.BUFFER },
-	{ number: FMT_NUMBER.BIGINT, bytes: FMT_BYTES.BUFFER },
+	{ number: FMT_NUMBER.NUMBER, bytes: FMT_BYTES.UINT8ARRAY },
+	{ number: FMT_NUMBER.HEX, bytes: FMT_BYTES.UINT8ARRAY },
+	{ number: FMT_NUMBER.STR, bytes: FMT_BYTES.UINT8ARRAY },
+	{ number: FMT_NUMBER.BIGINT, bytes: FMT_BYTES.UINT8ARRAY },
 
 	{ number: FMT_NUMBER.NUMBER, bytes: FMT_BYTES.UINT8ARRAY },
 	{ number: FMT_NUMBER.HEX, bytes: FMT_BYTES.UINT8ARRAY },